Server/server_requirements.txt
accelerate==0.32.1
huggingface-hub==0.24.0
Pillow==10.4.0
torch~=2.5.1+cu124
torchvision~=0.20.1+cu124
transformers==4.44.0
einops==0.8.0
gradio~=4.44.1
numpy~=1.26.4
tokenizers~=0.19.1
packaging~=24.2
datasets~=3.2.0
tqdm~=4.67.1
safetensors~=0.4.5
bcrypt~=4.2.1
dotenv~=0.9.9
python-dotenv~=1.0.1
opencv-python~=4.11.0.86